+++ IsekaiMod/Content/Features/ExceptionalFeats/ExceptionalFeatSelection.cs
@@ -3,6 +3,7 @@ using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Designers.Mechanics.Recommendations;
 using TabletopTweaks.Core.Utilities;
+using UnityEngine;
 using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Features.ExceptionalFeats {
@@ -11,6 +12,8 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
         private static readonly BlueprintFeatureSelection MythicFeatSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("9ee0f6745f555484299b0a1563b99d81");
         private static readonly BlueprintFeatureSelection ExtraFeatMythicFeat = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("e10c4f18a6c8b4342afe6954bde0587b");
         private static readonly BlueprintFeatureSelection ExtraMythicAbilityMythicFeat = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("8a6a511c55e67d04db328cc49aaad2b8");
+
+        private static readonly Sprite Icon_ExceptionalFeat = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_EXCEPTIONAL_FEAT.png");
 
         public static void Add() {
             // Add more Exceptional feats later
@@ -24,6 +27,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
                 bp.SetDescription(IsekaiContext, "Exceptional feats are feats that no ordinary NPC possess.\nSource: Isekai Mod");
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
+                bp.m_Icon = Icon_ExceptionalFeat;
                 bp.AddComponent<PureRecommendation>(c => {
                     c.Priority = RecommendationPriority.Good;
                 });
@@ -34,6 +38,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
                 bp.SetDescription(IsekaiContext, "Exceptional feats are feats that no ordinary NPC possess.\nSource: Isekai Mod");
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
+                bp.m_Icon = Icon_ExceptionalFeat;
                 bp.AddComponent<PureRecommendation>(c => {
                     c.Priority = RecommendationPriority.Good;
                 });